FoodTracker is an iOS meal tracking app.
The user can add, remove, or edit meals. Functions of the app include: adding a name for the meal, adding a picture from the user's photo library, and giving the meal a custom star rating. FoodTracker also comes with accessibility features including: VoiceOver, Guided Access, Text to Speech, and more.